Rewrite audio handling and unify with video frame queue (#9)
Audio handling is now simplified, audio frames are now handled through the existing video frame queue. Audio data was previously sent out via a callback, but is now sent as audio frames.

Audio data is only processed as required, in a similar way to video frames. This allows use of a filter pipeline with audio frames, which is currently implemented to provide audio stats per frame.